chore: repackage the Model Explorer UI
