+++ app/model-ui/src/main/java/org/trainbeans/model/ui/explorer/ElementChildFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trainbeans.model.ui;
+package org.trainbeans.model.ui.explorer;
 
 import java.beans.IntrospectionException;
 import java.util.List;
