+++ app/model-ui/src/main/java/org/trainbeans/model/ui/explorer/ModelTopComponent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trainbeans.model.ui;
+package org.trainbeans.model.ui.explorer;
 
 import java.awt.BorderLayout;
 import org.openide.awt.ActionID;
