+++ app/model-ui/src/test/java/org/trainbeans/model/ui/explorer/ElementClassChildFactoryTest.java
@@ -1,3 +1,4 @@
+
 /*
  * Copyright 2020 rhwood.
  *
@@ -13,15 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trainbeans.model.ui;
+package org.trainbeans.model.ui.explorer;
 
+import org.trainbeans.model.ui.explorer.ElementClassNode;
+import org.trainbeans.model.ui.explorer.ElementClassChildFactory;
 import java.util.ArrayList;
 import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openide.util.lookup.Lookups;
-import org.trainbeans.model.api.Model;
+import org.trainbeans.model.api.Element;
 import org.trainbeans.model.api.Turnout;
 import org.trainbeans.model.impl.DefaultModel;
 import org.trainbeans.model.impl.TurnoutFactory;
@@ -30,9 +34,9 @@ import org.trainbeans.model.impl.TurnoutFactory;
  *
  * @author rhwood
  */
-class ModelChildFactoryTest {
+class ElementClassChildFactoryTest {
 
-    private ModelChildFactory factory;
+    private ElementClassChildFactory factory;
     private DefaultModel model;
 
     @BeforeEach
@@ -40,20 +44,26 @@ class ModelChildFactoryTest {
         model = new DefaultModel(Lookups.fixed(new TurnoutFactory()));
         model.create(Turnout.class, "foo");
         model.create(Turnout.class, "bar");
-        factory = new ModelChildFactory();
+        factory = new ElementClassChildFactory(model);
     }
 
     @Test
     void testCreateKeys() {
-        List<Model> list = new ArrayList<>();
-        assertThat(list).isEmpty();
+        List<Class<? extends Element>> list = new ArrayList<>();
         assertThat(factory.createKeys(list)).isTrue();
-        assertThat(list).isNotEmpty();
+        assertThat(list)
+                .containsExactlyInAnyOrder(model.getCreatableClasses()
+                        .toArray(new Class[0]));
     }
 
     @Test
     void testCreateNodeForKey() {
-        assertThat(factory.createNodeForKey(model)).isNotNull();
+        assertThat(factory.createNodeForKey(Turnout.class))
+                .isNotNull()
+                .isInstanceOf(ElementClassNode.class);
+        assertThatCode(()
+                -> factory.createNodeForKey(null))
+                .isExactlyInstanceOf(NullPointerException.class);
     }
 
 }
